implement X and StringX
