--- XDateTimeService.cs
@@ -1,6 +1,0 @@
-﻿namespace Xes
-{
-    internal class XDateTimeService
-    {
-    }
-}
