+++ Xes.sln
@@ -2,7 +2,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.12.35707.178
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xes", "Xes.csproj", "{D17AF5EB-5089-4A69-8A50-C223DA408C62}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{4A5AE89B-196B-4C23-AECD-91D18DDFB11D}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xes", "src\Xes.csproj", "{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{631F7B4B-FA82-49BF-AA81-83AA02563EE5}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{BCEC094F-14EA-40CB-8FB5-2BFFB1D98836}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,12 +19,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D17AF5EB-5089-4A69-8A50-C223DA408C62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D17AF5EB-5089-4A69-8A50-C223DA408C62}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D17AF5EB-5089-4A69-8A50-C223DA408C62}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D17AF5EB-5089-4A69-8A50-C223DA408C62}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{CF06E7B6-96B3-4A6B-9184-6DF0791F8C7F} = {631F7B4B-FA82-49BF-AA81-83AA02563EE5}
 	EndGlobalSection
 EndGlobal
