+++ src/Coming/CurrencyX.cs
@@ -1,0 +1,5 @@
+﻿namespace Xes.Coming;
+
+internal class CurrencyX
+{
+}
