+++ src/Coming/Dates/DateOnlyX.cs
@@ -1,0 +1,5 @@
+﻿namespace Xes.Coming.Dates;
+
+internal class DateOnlyX
+{
+}
