+++ src/Coming/Dates/IXDateTimeService.cs
@@ -1,0 +1,5 @@
+﻿namespace Xes.Coming.Dates;
+
+internal interface IXDateTimeService
+{
+}
